// frontend/src/api/models/MainCustomQuest.ts
export interface MainCustomQuest {
    /**
     * 
     * @type {boolean}
     * @memberof MainCustomQuest
     */
    completed?: boolean;
    /**
     * 
     * @type {string}
     * @memberof MainCustomQuest
     */
    endDate?: string;
    /**
     * 
     * @type {number}
     * @memberof MainCustomQuest
     */
    experience?: number;
    /**
     * 
     * @type {number}
     * @memberof MainCustomQuest
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof MainCustomQuest
     */
    name?: string;
    /**
     * 
     * @type {Array<MainQuestsSkills>}
     * @memberof MainCustomQuest
     */
    skills?: Array<MainQuestsSkills>;
    /**
     * 
     * @type {string}
     * @memberof MainCustomQuest
     */
    startDate?: string;
    /**
     * 
     * @type {Array<MainRepeatingQuest>}
     * @memberof MainCustomQuest
     */
    subQuests?: Array<MainRepeatingQuest>;
    /**
     * 
     * @type {number}
     * @memberof MainCustomQuest
     */
    subQuestsCompleted?: number;
    /**
     * 
     * @type {string}
     * @memberof MainCustomQuest
     */
    type: string;
}